pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "birding-vibing"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.0.0",
    "folium>=0.18.0",
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.0.0",
    "streamlit>=1.51.0",
    "streamlit-folium>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
